Removed application of versioning.gradle from from root project, plus a little cleanup.
By removing the application of the versioning script from the root
project, all sub-projects are empowered to dynamically manage discrete
versioning information. When versioning.gradle is applied in the
"allprojects" configuration sub-projects don't pick up the timestamp
property. However, when applied in the "subprojects" configuration
each sub-project manages a discrete timestamp file in the build
directory.

I formatted the source code of the versioning.gradle file and that
caused a lot of noise. There was only a tiny bit of actual content
change in this file--see the assignment of ext.rcNumber.