Added new showBuildDir task
This new task will render a recursive listing of the build directory. This new
task is setup to depend on the 'build' task.

Also, I fixed an issue where clean is deleting the timestamp file used for
snapshot versions by moving it outside of the build directory. I also added
an entry to the .gitignore file so that this file won't look like it needs
to be added. The issue being addressed here is subtle and it's related to
the fact that versioning is established in the configuration phase and
clean does the actual cleaning in the proceeding execution phase.